+++ src/OrleansRuntime/OrleansRuntime.csproj
@@ -164,7 +164,9 @@
     <Compile Include="Silo\SiloControl.cs" />
     <Compile Include="Storage\MemoryStorageGrain.cs" />
     <Compile Include="Storage\StorageProviderManager.cs" />
+    <Compile Include="Streams\AdHoc\ObservableGrainExtension.cs" />
     <Compile Include="Streams\AdHoc\ObservableSupport.cs" />
+    <Compile Include="Streams\AdHoc\ObserverGrainExtension.cs" />
     <Compile Include="Streams\PersistentStream\PersistentStreamPullingManager.cs" />
     <Compile Include="Streams\PersistentStream\PersistentStreamPullingAgent.cs" />
     <Compile Include="Streams\PubSub\PubSubRendezvousGrain.cs" />
